Refactor: move api functions to /api
